requirements.txt
aiogram==3.5.0
tortoise-orm[asyncmy]
aerich
requests
yookassa
aioschedule
redis